rename the hull scripts
