+++ cms_dir/plot_convex_hull_quaternary.py
