+++ cms_dir/select_final_structures_ternary.py
